refactor(PayPalPaymentMethod): rename module to PayPalSavedPaymentMethod
Rename module and package (com.braintreepayments.api.paypalpaymentmethod ->
paypalsavedpaymentmethod) to align with the naming direction for the
saved-payment-method / Edit FI feature. Updates settings.gradle, root
build.gradle (source paths and detekt input), docs/package-list, the
Maven artifact id/pom name, and the release/release_snapshot workflow
job names and ids.
