--- PayPalPaymentMethod/src/main/res/values/strings.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="ppm_paypal_label">PayPal</string>
-    <string name="ppm_fi_card_masked_number">••%1$s</string>
-</resources>
