+++ PayPalSavedPaymentMethod/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    namespace "com.braintreepayments.api.paypalpaymentmethod"
+    namespace "com.braintreepayments.api.paypalsavedpaymentmethod"
     compileSdk rootProject.compileSdkVersion
 
     defaultConfig {
@@ -81,11 +81,11 @@ dependencies {
 
 // region signing and publishing
 
-project.ext.name = "paypal-payment-method"
-project.ext.pom_name = "paypal-payment-method"
+project.ext.name = "paypal-saved-payment-method"
+project.ext.pom_name = "paypal-saved-payment-method"
 project.ext.group_id = "com.braintreepayments.api"
 project.ext.version = rootProject.version
-project.ext.pom_desc = "PayPal Payment Method dependencies for Braintree Android."
+project.ext.pom_desc = "PayPal Saved Payment Method dependencies for Braintree Android."
 
 apply from: rootProject.file("gradle/gradle-publish.gradle")
