+++ PayPalSavedPaymentMethod/src/androidTest/java/com/braintreepayments/api/paypalsavedpaymentmethod/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.paypalpaymentmethod
+package com.braintreepayments.api.paypalsavedpaymentmethod
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.braintreepayments.api.paypalpaymentmethod.test", appContext.packageName)
+        assertEquals("com.braintreepayments.api.paypalsavedpaymentmethod.test", appContext.packageName)
     }
 }
