+++ PayPalSavedPaymentMethod/src/main/AndroidManifest.xml
