+++ PayPalSavedPaymentMethod/src/main/res/drawable/ic_edit_pencil.xml
