+++ PayPalSavedPaymentMethod/src/main/res/drawable/ic_fi_bank_placeholder.xml
