+++ PayPalSavedPaymentMethod/src/main/res/drawable/ic_fi_card_placeholder.xml
