+++ PayPalSavedPaymentMethod/src/main/res/drawable/ic_paypal_brand_logo.xml
