+++ PayPalSavedPaymentMethod/src/test/java/com/braintreepayments/api/paypalsavedpaymentmethod/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.paypalpaymentmethod
+package com.braintreepayments.api.paypalsavedpaymentmethod
 
 import org.junit.Test
