+++ docs/package-list
@@ -15,8 +15,8 @@ module:LocalPayment
 com.braintreepayments.api.localpayment
 module:PayPal
 com.braintreepayments.api.paypal
-module:PayPalPaymentMethod
-com.braintreepayments.api.paypalpaymentmethod
+module:PayPalSavedPaymentMethod
+com.braintreepayments.api.paypalsavedpaymentmethod
 module:SEPADirectDebit
 com.braintreepayments.api.sepadirectdebit
 module:ShopperInsights
