+++ settings.gradle
@@ -28,6 +28,6 @@ include ':DataCollector'
 include ':PayPalMessaging'
 include ':ShopperInsights'
 include ':UIComponents'
-include ':PayPalPaymentMethod'
+include ':PayPalSavedPaymentMethod'
 
 includeBuild 'DokkaRestrictToPlugin'
